split Character Showing core into base class
